Update existing tests.
